renamed dashboard to status bar, added background color
